Transcripts - Style cues into paragraphs (#2468)

+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/podcast/TranscriptsManagerImpl.kt
@@ -14,10 +14,10 @@ class TranscriptsManagerImpl @Inject constructor(
         episodeUuid: String,
         transcripts: List<Transcript>,
     ) {
+        if (transcripts.isEmpty()) return
+        transcriptDao.deleteForEpisode(episodeUuid)
         findBestTranscript(transcripts)?.let { bestTranscript ->
             transcriptDao.insert(bestTranscript)
-        } ?: run {
-            transcriptDao.deleteForEpisode(episodeUuid)
         }
     }
